feat(ContentView): Separate Chip Component
