--- AllerView/Component.swift
@@ -1,8 +1,0 @@
-//
-//  Component.swift
-//  AllerView
-//
-//  Created by Hyemi on 2023/07/16.
-//
-
-import Foundation
